+++ src/main/java/com/semosan/api/domain/user/dto/command/CreateUserOnboardingCommand.java
@@ -1,14 +1,14 @@
 package com.semosan.api.domain.user.dto.command;
 
 import com.semosan.api.domain.user.entity.User;
-import com.semosan.api.domain.user.enums.ExerciseDuration;
-import com.semosan.api.domain.user.enums.ExerciseFrequency;
-import com.semosan.api.domain.user.enums.ExerciseType;
-import com.semosan.api.domain.user.enums.FitnessLevel;
-import com.semosan.api.domain.user.enums.HikingGoalType;
-import com.semosan.api.domain.user.enums.HikingLevel;
-import com.semosan.api.domain.user.enums.HikingPurpose;
-import com.semosan.api.domain.user.enums.PreferredDifficulty;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseDuration;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseFrequency;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseType;
+import com.semosan.api.domain.user.enums.onboarding.FitnessLevel;
+import com.semosan.api.domain.user.enums.onboarding.HikingGoalType;
+import com.semosan.api.domain.user.enums.onboarding.HikingLevel;
+import com.semosan.api.domain.user.enums.onboarding.HikingPurpose;
+import com.semosan.api.domain.user.enums.onboarding.PreferredDifficulty;
 
 public record CreateUserOnboardingCommand(
         User user,
