+++ src/main/java/com/semosan/api/domain/user/dto/request/RegisterOnboardingRequest.java
@@ -1,13 +1,13 @@
 package com.semosan.api.domain.user.dto.request;
 
-import com.semosan.api.domain.user.enums.ExerciseDuration;
-import com.semosan.api.domain.user.enums.ExerciseFrequency;
-import com.semosan.api.domain.user.enums.ExerciseType;
-import com.semosan.api.domain.user.enums.Gender;
-import com.semosan.api.domain.user.enums.HikingGoalType;
-import com.semosan.api.domain.user.enums.HikingLevel;
-import com.semosan.api.domain.user.enums.HikingPurpose;
-import com.semosan.api.domain.user.enums.PreferredDifficulty;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseDuration;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseFrequency;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseType;
+import com.semosan.api.domain.user.enums.user.Gender;
+import com.semosan.api.domain.user.enums.onboarding.HikingGoalType;
+import com.semosan.api.domain.user.enums.onboarding.HikingLevel;
+import com.semosan.api.domain.user.enums.onboarding.HikingPurpose;
+import com.semosan.api.domain.user.enums.onboarding.PreferredDifficulty;
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
