+++ src/main/java/com/semosan/api/domain/user/entity/User.java
@@ -3,10 +3,10 @@ package com.semosan.api.domain.user.entity;
 import com.semosan.api.common.base.BaseEntity;
 import com.semosan.api.domain.user.dto.command.CompleteOnboardingCommand;
 import com.semosan.api.domain.user.dto.command.UpdateUserProfileCommand;
-import com.semosan.api.domain.user.enums.DeviceType;
-import com.semosan.api.domain.user.enums.Gender;
-import com.semosan.api.domain.user.enums.OAuthProvider;
-import com.semosan.api.domain.user.enums.OnboardingStatus;
+import com.semosan.api.domain.user.enums.user.DeviceType;
+import com.semosan.api.domain.user.enums.user.Gender;
+import com.semosan.api.domain.user.enums.user.OAuthProvider;
+import com.semosan.api.domain.user.enums.user.OnboardingStatus;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.JdbcTypeCode;
