+++ src/main/java/com/semosan/api/domain/user/entity/UserOnboarding.java
@@ -2,14 +2,14 @@ package com.semosan.api.domain.user.entity;
 
 import com.semosan.api.common.base.BaseEntity;
 import com.semosan.api.domain.user.dto.command.CreateUserOnboardingCommand;
-import com.semosan.api.domain.user.enums.ExerciseDuration;
-import com.semosan.api.domain.user.enums.ExerciseFrequency;
-import com.semosan.api.domain.user.enums.ExerciseType;
-import com.semosan.api.domain.user.enums.FitnessLevel;
-import com.semosan.api.domain.user.enums.HikingGoalType;
-import com.semosan.api.domain.user.enums.HikingLevel;
-import com.semosan.api.domain.user.enums.HikingPurpose;
-import com.semosan.api.domain.user.enums.PreferredDifficulty;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseDuration;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseFrequency;
+import com.semosan.api.domain.user.enums.onboarding.ExerciseType;
+import com.semosan.api.domain.user.enums.onboarding.FitnessLevel;
+import com.semosan.api.domain.user.enums.onboarding.HikingGoalType;
+import com.semosan.api.domain.user.enums.onboarding.HikingLevel;
+import com.semosan.api.domain.user.enums.onboarding.HikingPurpose;
+import com.semosan.api.domain.user.enums.onboarding.PreferredDifficulty;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
