+++ src/main/java/com/semosan/api/domain/user/enums/onboarding/ExerciseDuration.java
@@ -1,4 +1,4 @@
-package com.semosan.api.domain.user.enums;
+package com.semosan.api.domain.user.enums.onboarding;
 
 public enum ExerciseDuration {
     OVER_4H,
