+++ src/main/java/com/semosan/api/domain/user/enums/user/OAuthProvider.java
@@ -1,4 +1,4 @@
-package com.semosan.api.domain.user.enums;
+package com.semosan.api.domain.user.enums.user;
 
 public enum OAuthProvider {
     KAKAO,
