+++ src/main/java/com/semosan/api/domain/user/enums/user/OnboardingStatus.java
@@ -1,4 +1,4 @@
-package com.semosan.api.domain.user.enums;
+package com.semosan.api.domain.user.enums.user;
 
 public enum OnboardingStatus {
     INCOMPLETE,  // 온보딩 미완료
