+++ src/main/java/com/semosan/api/domain/user/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.semosan.api.domain.user.repository;
 
 import com.semosan.api.domain.user.entity.User;
-import com.semosan.api.domain.user.enums.OAuthProvider;
+import com.semosan.api.domain.user.enums.user.OAuthProvider;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
