+++ src/main/java/com/semosan/api/domain/user/service/UserService.java
@@ -5,8 +5,8 @@ import com.semosan.api.common.status.ErrorStatus;
 import com.semosan.api.domain.user.dto.command.UpdateUserProfileCommand;
 import com.semosan.api.domain.user.dto.request.UpdateUserProfileRequest;
 import com.semosan.api.domain.user.entity.User;
-import com.semosan.api.domain.user.enums.DeviceType;
-import com.semosan.api.domain.user.enums.OAuthProvider;
+import com.semosan.api.domain.user.enums.user.DeviceType;
+import com.semosan.api.domain.user.enums.user.OAuthProvider;
 import com.semosan.api.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
